Health Checks + MongoDB Connection Cleanup
 - Added Health Check endpoint
 - Moved MongoDB Connectiont to server.js
 - Use env vars for mongoDB creds
