+++ app/models/message.model.js
@@ -1,11 +1,5 @@
 const mongoose = require('mongoose')
-const mongoURL = 'mongodb+srv://PalindromeUser:123@mongocluster1.1b0kw.mongodb.net/PalindromesDB?retryWrites=true&w=majority'
-const mongoOptions = { 
-	useNewUrlParser: true, 
-	useUnifiedTopology: true, 
-	useCreateIndex: true 
-}
-mongoose.connect(mongoURL, mongoOptions);
+
 
 //define schema
 const MessageSchema = new mongoose.Schema({
