+++ app/routes/health.server.routes.js
@@ -1,0 +1,10 @@
+'use strict';
+
+let express = require('express')
+let router = express.Router()
+
+const health = require('../controllers/health.server.controller');
+router.get('/health', health.getHealth)
+
+
+module.exports = router
